CDH-81317: fix Impala container Kudu client build
This is required for the Impala container build to locate the Kudu
client library. A corresponding fix in the Impala code is also required
to solve CDH-81317, but this change can be merged ahead of that.

The critical variable is KUDU_CLIENT_DIR, but passing all of the same
configuration will head off other potential issues.

Change-Id: I4d1da24e8629b38750a2aa191bb3d8d27dda6a2a