CDPD-3137: Build Impala with a Kudu built with the native toolchain
Currently, Impala is building against a mix of CDP and CDH versions
of Kudu. The CDP build passes in KUDU_CLIENT_DIR pointing to the CDP
version of Kudu, but the rest of Impala's build uses the hardcoded
CDH version of Kudu.

This switches Impala to building against CDP Kudu by adding a text
replace for IMPALA_KUDU_VERSION and IMPALA_KUDU_JAVA_VERSION. Impala
also needs to build against a version of Kudu built with its native
toolchain. This changes Impala's build steps to build Kudu with the
native toolchain and then pass it in as KUDU_CLIENT_DIR. It also
produces a Kudu-for-impala tarball artifact that is useful for
Impala's upstream tests.

Testing:
 - Ran a Canary build and verified the artifacts

Change-Id: I0a5acd5990c5e9a2a88222db26ca7b4e2e3e3b96