CDPD-4271 exclude maven project 'maven-plugin' from accumulo build
During CDPD-4271 we upgraded some Hadoop dependencies and were hitting a known maven bug
during accumulo compilation: https://issues.apache.org/jira/browse/MPLUGIN-312
As the affected accumulo module is not important for us, we decided to exclude it from the
build. Other solution would be to upgrade the currently used maven 3.3.9 in the build
infrastructure, but that can cause failures in other components.

Change-Id: Ie49178926d1a0ecbb4d62f66b68168d850642810