Kudu: ensure CSD artifact versions match the CDPD branch version
Currently all CSD artifacts are versioned as 7.x.0 because that value is
statically defined in Kudu. This yields CSD JAR names like KUDU-7.x.0.jar
regardless of the branch the CSD was built on. This patch adds build-time
STACKVERSION substitution for that value, which will produce names like
KUDU-7.0.2.0.jar when building cdpd-master.

That still doesn't quite match the versioning of CM's built-in CSDs (7.x.0
on cdpd-master and e.g. 7.0.1 on other branches), but my understanding is
that CM versioning will change in the near future to match CDPD versioning.

Change-Id: I2e331e6bff53c51f748c9ae1b552ef146aa48235