CDPD-5390. Change Hive dependency to version in CDP
When Livy 0.6 was merged into livy cdpd-master, a new dependency on Hive
was introduced. The hive.version used upstream is 3.0.0. Here, we replace
it by the Hive version in CDP.
There is no spark1.version in the livy pom.xml, so remove the replacement
of spark1.version at the same time.
Also, specify that livy depends on hadoop and hive.

Change-Id: Ia75762696cebf4760a81c4d9543c711190604440