CDPD-10444 (cdpd side): Update the version of Atlas' Impala hook
Due to the incompatibility of jackson-databind reported at CDPD-10218,
at the end of a query execution in Impala, we saw an error of
NoClassDefFoundError resulting from the call to
QueryEventHookManager#executeQueryCompleteHooks(). Now that the issue
has been resolved on the Atlas side, Impala should use an upgraded hook
provided by Atlas.

To address this issue, instead of using a hook of a hard-coded version
that has the fix as previously done, this patch enables Impala's build
script to always download the up-to-date hook. Specifically, we instruct
components.ini to download the tarball of the latest hook to a
temporary directory accessible by Impala's build script, allowing the
build script to include the latest hook in Impala's docker images in the
CDP build.

Testing:
- Verified that this patch could be successfully built in the Canary
  tool with and without the fix on the Impala side.

Change-Id: I5c673f1875343aaf08b9e644d59450ef48624b69